Update k6 and add tests around races
Before the update of k6 it was possible that a websocket connection will start
stopping due to context Done. That will use the eventloop to synchronise
the update of the statuses and the closing of the done channel.

And everything will work *except* if the eventloop was not already
stopping due to unhandled exception (for example).

In that case the eventloop will just not run the call to stop the
connection.

Detecting that is inherently hard and will require to then have
a separate mechanism to close and stop the websocket connection that is
only used in this particular case.

The test provided also caught leakage of goroutines which is now caught
and fixed.

fixes #36